weather-project/requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
datasets
pillow
onnx